+++ lib/src/replication/concurrency-controller/create-mutex-concurrency-controller.ts
@@ -2,7 +2,7 @@ import { Mutex } from 'async-mutex';
 import { ReplicationConcurrencyController } from './concurrency-controller';
 
 /**
- * Uses a single mutex to execute a single message at a time in order
+ * Uses a single mutex to execute a single message at a time in the WAL sort order
  * @returns The controller to acquire and release the mutex
  */
 export const createReplicationMutexConcurrencyController =